Move module contents out of mod.rs
